Extract commons from SubmitCommand
